Time: 0 ms (100%), Space: 8.9 MB (1.57%) - LeetHub

+++ 0441-arranging-coins/0441-arranging-coins.cpp
@@ -1,9 +1,9 @@
 class Solution {
 public:
     int arrangeCoins(int n) {
-        long long low=0;
+        int low=0;
 
-        long long high=n;
+        int high=n;
 
         while(low<=high){
